+++ backend/src/models/post.go
@@ -2,7 +2,7 @@ package models
 
 type Post struct {
 	Model
-	CategoryId      int16  `json:"category_id"`
+	CategoryId      int    `json:"category_id"`
 	Title           string `json:"title"`
 	Contents        string `json:"contents"`
 	PreviewContents string `json:"preview_contents"`
